requirements.txt
discord.py >=1.3.3
google-assistant-sdk[samples] >= 0.6.0
dblpy >= 0.3.4
html2text >= 2020.1.16